move get usdc token to account service
